Start making the code a bit prettier (sectioned)
