+++ src/blocks.h
@@ -1,0 +1,22 @@
+/*
+
+    Blocks header
+    Make 12/6/2025 @ 6:01PM by Digi-Space Productions (EGAMatsu)
+
+*/
+
+#ifndef BLOCKS_H_
+#define BLOCKS_H_
+
+enum { 
+    BLK_GRASS=0, BLK_DIRT, BLK_STONE, BLK_WATER, BLK_WOOD, BLK_LEAF, 
+    BLK_SAND, BLK_BRICK, BLK_PLANK, BLK_COBBLE, BLK_BEDROCK,
+    BLK_OBSIDIAN, BLK_NETHERRACK, BLK_PORTAL, BLK_LAVA,
+    BLK_GRASS_SIDE, 
+    BLK_COUNT 
+};
+
+#define ITEM_FLINT_STEEL 100 
+#define SHADES_PER_BLOCK 8
+
+#endif
