+++ src/main.c
@@ -10,365 +10,8 @@ Last Edited 12/6/2025
 
 */
 
-
-
-
-#include <gba.h>
-#include <stdlib.h>
-#include <string.h>
-#include <stdint.h>
-#include <stdbool.h>
-#include <math.h>
-#include <stdarg.h>
-#include <stdio.h>
-
-
-#include "image0.c"
-#include "image1.c"
-
-extern const u8 image0[];
-extern const u8 image1[];
-
-
-typedef uint8_t  u8;
-typedef uint16_t u16;
-typedef uint32_t u32;
-typedef int32_t  fix;
-
-typedef struct { 
-    bool start_from_load; 
-    int  load_slot; 
-} MenuResult;
-
-typedef enum { 
-    PM_RESUME, 
-    PM_TO_MAINMENU 
-} PauseResult;
-
-typedef enum { 
-    WORLD_CLASSIC, 
-    WORLD_SUPERFLAT 
-} WorldType;
-
-
-#undef IWRAM_CODE
-#define IWRAM_CODE __attribute__((section(".iwram"), noinline))
-#define ARM_CODE   __attribute__((target("arm")))
-
-#ifndef EWRAM_DATA
-#define EWRAM_DATA __attribute__((section(".ewram")))
-#endif
-#ifndef IWRAM_DATA
-#define IWRAM_DATA __attribute__((section(".iwram")))
-#endif
-
-#define W 240
-#define H 160
-
-#define VRAM_PAGE0 ((volatile u8*)0x06000000)
-#define VRAM_PAGE1 ((volatile u8*)0x0600A000)
-#define PAL_BG_MEM ((volatile u16*)0x05000000)
-
-
-#define PRECISION 16
-#define FONE   ((fix)(1 << PRECISION))
-#define F(x)   ((fix)((x) * (1 << PRECISION)))
-#define I2F(i) ((fix)((i) << PRECISION))
-#define F2I(x) ((int)((x) >> PRECISION))
-
-static inline fix fmuli(fix a, fix b){ return (fix)(((int32_t)(a>>8)*(b>>8))); }
-
-static inline fix fdivi(fix a, fix b){
-    if(b == 0) return (a < 0) ? -FONE*1000 : FONE*1000;
-    return (fix)(Div((int32_t)a<<8, b)<<8);
-}
-
-
-static const char GBA_SAVE_TYPE[] __attribute__((used)) = "SRAM_V110";
-static inline void keep_save_id(void){ __asm__ __volatile__("" :: "r"(GBA_SAVE_TYPE)); }
-
-
-static volatile u8* backbuffer  = VRAM_PAGE1;
-static inline void vblank(void) { VBlankIntrWait(); }
-static inline void flip(void) {
-    REG_DISPCNT ^= BACKBUFFER;
-    backbuffer = (volatile u8*)((REG_DISPCNT & BACKBUFFER) ? VRAM_PAGE0 : VRAM_PAGE1);
-}
-
-
-static inline fix flerp(fix a, fix b, fix t){ return a + fmuli((b-a), t); }
-static inline fix fixabs(fix x){ return (x < 0) ? -x : x; }
-
-fix IWRAM_CODE ARM_CODE fsqrt(fix n) {
-    if (n <= 0) return 0;
-    fix root = n, last;
-    int iter = 0;
-    do { 
-        last = root; 
-        root = (root + fdivi(n, root)) >> 1; 
-        iter++;
-    } while (root != last && iter < 20);
-    return root;
-}
-
-typedef struct { fix x,y,z; } v3;
-typedef struct { int x,y; fix u,v; } v2uv;
-
-
-#define LUT_N 512
-static EWRAM_DATA fix sinLUT[LUT_N];
-static inline fix fsin(int a){ return sinLUT[a & (LUT_N-1)]; }
-static inline fix fcos(int a){ return sinLUT[(a + (LUT_N>>2)) & (LUT_N-1)]; }
-
-#define INVZ_LUT_SIZE 256
-#define INVZ_NEAR_FIX F(0.25f)
-#define INVZ_FAR_FIX  F(40.0f)
-static IWRAM_DATA fix invz_lut[INVZ_LUT_SIZE];
-static IWRAM_DATA fix reciprocal_lut[256];
-
-static void build_luts(void){
-    for(int i=0; i<LUT_N; i++){
-        sinLUT[i] = F(sinf((float)i * 2.0f * 3.14159265f / (float)LUT_N));
-    }
-    for(int i=0; i<INVZ_LUT_SIZE; i++){
-        fix z = INVZ_NEAR_FIX + ((INVZ_FAR_FIX - INVZ_NEAR_FIX) * i) / INVZ_LUT_SIZE;
-        invz_lut[i] = fdivi(FONE, z);
-    }
-    reciprocal_lut[0] = FONE;
-    for(int i=1;i<256;i++) reciprocal_lut[i] = fdivi(FONE, I2F(i));
-}
-
-static IWRAM_CODE fix fast_invz(fix z){
-    if(z <= INVZ_NEAR_FIX) return invz_lut[0];
-    if(z >= INVZ_FAR_FIX)  return invz_lut[INVZ_LUT_SIZE-1];
-    fix range = INVZ_FAR_FIX - INVZ_NEAR_FIX;
-    fix normalized = fdivi(z - INVZ_NEAR_FIX, range);
-    int idx = F2I(fmuli(normalized, I2F(INVZ_LUT_SIZE-1)));
-    if(idx < 0) idx = 0;
-    if(idx >= INVZ_LUT_SIZE-1) idx = INVZ_LUT_SIZE-1;
-    return invz_lut[idx];
-}
-
-
-enum { 
-    BLK_GRASS=0, BLK_DIRT, BLK_STONE, BLK_WATER, BLK_WOOD, BLK_LEAF, 
-    BLK_SAND, BLK_BRICK, BLK_PLANK, BLK_COBBLE, BLK_BEDROCK,
-    BLK_OBSIDIAN, BLK_NETHERRACK, BLK_PORTAL, BLK_LAVA,
-    BLK_GRASS_SIDE, 
-    BLK_COUNT 
-};
-
-#define ITEM_FLINT_STEEL 100 
-#define SHADES_PER_BLOCK 8
-
-
-static const u16 palBaseColors[BLK_COUNT] = {
-    RGB5(3, 13, 3),   
-    RGB5(14, 9, 4),   
-    RGB5(16, 16, 16), 
-    RGB5(4, 8, 24),   
-    RGB5(16, 11, 6),  
-    RGB5(4, 18, 4),   
-    RGB5(26, 25, 15), 
-    RGB5(20, 8, 6),   
-    RGB5(22, 16, 8),  
-    RGB5(12, 12, 12), 
-    RGB5(5, 5, 5),    
-    RGB5(6, 4, 10),   
-    RGB5(18, 6, 6),   
-    RGB5(28, 5, 31),  
-    RGB5(31, 15, 0),  
-RGB5(10, 7, 4) 
-};
-
-
-static EWRAM_DATA u8 textures[BLK_COUNT][64];
-
-static const u8 texture_data_rom[BLK_COUNT][64] = {
-    
-    {
-        3,3,2,3,3,2,3,3,  2,3,4,2,3,3,2,4, 3,2,3,2,3,3,2,3,  3,4,2,3,3,2,3,2,
-        2,3,2,4,2,3,3,2,  4,2,3,2,3,3,2,3, 3,2,3,3,3,2,2,3,  3,3,2,4,2,3,3,2
-    },
-    
-    {
-        2,3,1,4,2,3,1,3,  3,1,3,2,4,1,3,2, 1,4,2,3,1,3,2,3,  3,2,3,1,3,4,1,3,
-        3,1,3,2,3,4,1,2,  4,3,1,3,2,3,2,1, 2,3,4,1,3,2,3,4,  1,3,2,3,4,1,3,2
-    },
-    
-    {
-        2,1,3,2,1,3,2,1,  3,2,1,2,3,1,2,3, 1,3,2,3,1,2,3,2,  2,1,3,1,2,3,1,2,
-        3,2,1,2,3,2,1,3,  1,3,2,3,1,2,3,1, 2,1,3,1,2,3,2,1,  3,2,1,2,3,1,2,3
-    },
-    
-    {
-        3,3,1,1,3,3,1,1,  1,3,3,1,1,3,3,1, 1,1,3,3,1,1,3,3,  3,1,1,3,3,1,1,3,
-        3,3,1,1,3,3,1,1,  1,3,3,1,1,3,3,1, 1,1,3,3,1,1,3,3,  3,1,1,3,3,1,1,3
-    },
-    
-    {
-        1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2, 1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2,
-        1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2, 1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2
-    },
-    
-    {
-        2,2,2,3,2,2,1,2,  2,1,2,2,2,3,2,2, 2,3,2,2,2,1,2,2,  2,2,2,2,2,2,2,1,
-        1,2,2,2,3,2,2,2,  2,2,1,2,2,2,3,2, 2,2,3,2,2,2,1,2,  3,2,2,2,1,2,2,2
-    },
-    
-    {
-        3,2,3,4,2,3,3,2,  3,3,2,3,4,2,3,3, 2,3,4,3,2,3,3,2,  4,2,3,3,2,3,4,2,
-        3,4,2,3,3,2,3,4,  2,3,3,2,3,3,2,3, 3,2,3,3,4,2,3,3,  3,4,2,3,3,2,3,2
-    },
-    
-    {
-        3,3,3,0,3,3,3,3,  3,3,3,0,3,3,3,3, 3,3,3,0,3,3,3,3,  0,0,0,0,0,0,0,0,
-        3,3,3,3,3,3,3,0,  3,3,3,3,3,3,3,0, 3,3,3,3,3,3,3,0,  0,0,0,0,0,0,0,0
-    },
-    
-    {
-        1,1,1,1,1,1,1,1,  2,1,2,2,1,2,2,1, 2,2,1,2,2,1,2,2,  1,1,1,1,1,1,1,1,
-        1,2,2,1,2,2,1,2,  2,1,2,2,1,2,2,1, 2,2,1,2,2,1,2,2,  1,1,1,1,1,1,1,1
-    },
-    
-    {
-        1,1,3,3,3,1,1,2,  2,2,3,3,3,2,2,1, 1,3,3,3,1,1,2,2,  2,3,3,3,2,2,1,1,
-        3,3,3,1,1,2,2,3,  3,3,2,2,1,1,3,3, 1,1,1,2,2,3,3,3,  1,1,2,2,3,3,1,1
-    },
-    
-    {
-        0,4,1,5,0,3,2,5,  4,1,5,0,2,5,1,3, 5,2,0,3,5,1,4,0,  1,5,2,4,0,3,5,2,
-        0,3,5,1,4,0,2,5,  3,0,4,2,5,1,0,4, 4,1,2,5,0,3,2,1,  5,2,0,3,1,4,2,0
-    },
-    
-    {
-        1,2,1,1,1,2,1,1,  1,1,2,1,1,1,2,1, 2,1,1,1,2,1,1,1,  1,1,2,1,1,1,2,1,
-        1,2,1,1,1,2,1,1,  1,1,2,1,1,1,2,1, 2,1,1,1,2,1,1,1,  1,1,2,1,1,1,2,1
-    },
-    
-    {
-        2,1,3,2,1,3,2,1,  3,2,1,2,3,1,2,3, 1,3,2,3,1,2,3,2,  2,1,3,1,2,3,1,2,
-        3,2,1,2,3,2,1,3,  1,3,2,3,1,2,3,1, 2,1,3,1,2,3,2,1,  3,2,1,2,3,1,2,3
-    },
-    
-    {
-        2,3,4,3,2,3,4,3,  3,4,3,2,3,4,3,2, 4,3,2,3,4,3,2,3,  3,2,3,4,3,2,3,4,
-        2,3,4,3,2,3,4,3,  3,4,3,2,3,4,3,2, 4,3,2,3,4,3,2,3,  3,2,3,4,3,2,3,4
-    },
-    
-
-
-
-    {
-        1,4,6,4,1,4,6,4,  6,2,0,2,6,4,2,4, 4,0,2,0,4,6,1,6,  4,6,4,6,4,6,4,6,
-        1,4,6,4,1,4,6,4,  6,2,0,2,6,4,2,4, 4,0,2,0,4,6,1,6,  4,6,4,6,4,6,4,6
-    },
-
-
-
-
-
-
-
-
-
-
-   {
-        5,5,5,5,5,5,5,5, 5,0,5,0,5,5,0,5,  1,0,0,1,0,0,0,0, 0,0,0,0,1,0,1,0,
-        1,0,0,1,0,0,1,0, 0,0,0,0,0,1,0,0,  0,1,0,0,1,0,1,0, 1,0,0,1,0,0,0,1
-    }
-};
-
-static inline u16 shadeRGB15(u16 c, int light){
-    int r = (c & 31), g = ((c>>5) & 31), b = ((c>>10)&31);
-    int num = 4 + light * 2; 
-    int den = 16;
-    r = (r * num) / den; if(r>31) r=31;
-    g = (g * num) / den; if(g>31) g=31;
-    b = (b * num) / den; if(b>31) b=31;
-    return (u16)(r | (g<<5) | (b<<10));
-}
-
-#define SKY_START_IDX 128
-#define VOID_START_IDX 160
-#define UI_START_IDX 250
-#define SKY_COLORS 32
-
-static int g_dimension = 0; 
-
-static void buildPalette(void){
-    
-    for(int b=0; b<BLK_COUNT; b++){
-        for(int s=0; s<SHADES_PER_BLOCK; s++){
-        if (b == BLK_GRASS_SIDE) {
-                
-
-
-
-
-                if (s <= 4) {
-                    
-                    
-                    PAL_BG_MEM[b*8 + s] = shadeRGB15(RGB5(14, 9, 4), s*2 + 2); 
-                } else {
-                    
-                    
-                    PAL_BG_MEM[b*8 + s] = shadeRGB15(RGB5(3, 13, 3), (s - 5) * 4 + 4);
-                }
-            } else if (b == BLK_PORTAL) {
-                int bright = 12 + s*2; if(bright>31) bright=31;
-                PAL_BG_MEM[b*8 + s] = RGB5(bright, 6, bright); 
-            } else if (b == BLK_LAVA) {
-                
-                if (s <= 2) {
-                    PAL_BG_MEM[b*8 + s] = RGB5(25 + s*2, 4 + s*2, 0); 
-                } else {
-                    int r = 31;
-                    int g = 12 + (s-3)*4; if(g>31) g=31;
-                    PAL_BG_MEM[b*8 + s] = RGB5(r,g,0);
-                }
-            } else {
-                PAL_BG_MEM[b*8 + s] = shadeRGB15(palBaseColors[b], s*2);
-            }
-        }
-    }
-    
-    
-    if (g_dimension == 0) {
-        for(int i=0; i<SKY_COLORS; i++) {
-            int r = 31 - (21 * i) / SKY_COLORS;
-            int g = 31 - (13 * i) / SKY_COLORS;
-            int b = 31;
-            PAL_BG_MEM[SKY_START_IDX + i] = RGB5(r,g,b);
-        }
-        for(int i=0; i<SKY_COLORS; i++) {
-            int val = 31 - (26 * i) / SKY_COLORS;
-            PAL_BG_MEM[VOID_START_IDX + i] = RGB5(val, val, val);
-        }
-    } else {
-        for(int i=0; i<SKY_COLORS; i++) {
-            int r = 16 - (8 * i) / SKY_COLORS;
-            PAL_BG_MEM[SKY_START_IDX + i] = RGB5(r,3,3);
-        }
-        for(int i=0; i<SKY_COLORS; i++) {
-             int r = 10 - (8 * i) / SKY_COLORS;
-             PAL_BG_MEM[VOID_START_IDX + i] = RGB5(r, 0, 0);
-        }
-    }
-    
-    
-    PAL_BG_MEM[255] = RGB5(31,31,31); 
-    PAL_BG_MEM[254] = RGB5(31,31,0);  
-    PAL_BG_MEM[253] = RGB5(15,15,15); 
-    PAL_BG_MEM[252] = RGB5(8,8,8);    
-    PAL_BG_MEM[251] = RGB5(20,0,0);
-    PAL_BG_MEM[250] = RGB5(20,20,20); 
-}
-
-static void generate_textures(void){
-    memcpy(textures, texture_data_rom, sizeof(texture_data_rom));
-}
-
+#include "platform.h"
+int g_dimension = 0; 
 
 #define CW 80
 #define CH 32
@@ -379,16 +22,6 @@ static inline int widx(int x,int y,int z){
     if(x < 0 || x >= CW || y < 0 || y >= CH || z < 0 || z >= CD) return CW*CH*CD/2;
     return (y*CD + z)*CW + x; 
 }
-
-
-#define SRAM_BASE ((volatile u8*)0x0E000000)
-#define SLOT_SIZE 0x3C00       
-#define SLOT0_OFF 0x0000
-#define SLOT1_OFF 0x3C00       
-#define SLOT_HDR_PAD 64
-#define MAX_EDITS  3776        
-#define EDIT_AREA_SIZE (MAX_EDITS*4)  
-#define SLOT_CHKSUM_OFF (SLOT_SIZE - 4)
 
 typedef struct {
     u32 tag; u32 version; u32 seed;
