+++ src/textures.h
@@ -1,0 +1,117 @@
+/*
+
+    Textures header
+    Created 12/7/2025 @ 5:59 PM by Digi-Space Productions (EGAMatsu)
+
+*/
+
+#ifndef TEXTURES_H_
+#define TEXTURES_H_
+
+#include "platform.h"
+
+static const u16 palBaseColors[BLK_COUNT] = {
+    RGB5(3, 13, 3),   
+    RGB5(14, 9, 4),   
+    RGB5(16, 16, 16), 
+    RGB5(4, 8, 24),   
+    RGB5(16, 11, 6),  
+    RGB5(4, 18, 4),   
+    RGB5(26, 25, 15), 
+    RGB5(20, 8, 6),   
+    RGB5(22, 16, 8),  
+    RGB5(12, 12, 12), 
+    RGB5(5, 5, 5),    
+    RGB5(6, 4, 10),   
+    RGB5(18, 6, 6),   
+    RGB5(28, 5, 31),  
+    RGB5(31, 15, 0),  
+RGB5(10, 7, 4) 
+};
+
+static EWRAM_DATA u8 textures[BLK_COUNT][64];
+
+
+static const u8 texture_data_rom[BLK_COUNT][64] = {
+    {
+        3,3,2,3,3,2,3,3,  2,3,4,2,3,3,2,4, 3,2,3,2,3,3,2,3,  3,4,2,3,3,2,3,2,
+        2,3,2,4,2,3,3,2,  4,2,3,2,3,3,2,3, 3,2,3,3,3,2,2,3,  3,3,2,4,2,3,3,2
+    },
+    
+    {
+        2,3,1,4,2,3,1,3,  3,1,3,2,4,1,3,2, 1,4,2,3,1,3,2,3,  3,2,3,1,3,4,1,3,
+        3,1,3,2,3,4,1,2,  4,3,1,3,2,3,2,1, 2,3,4,1,3,2,3,4,  1,3,2,3,4,1,3,2
+    },
+    
+    {
+        2,1,3,2,1,3,2,1,  3,2,1,2,3,1,2,3, 1,3,2,3,1,2,3,2,  2,1,3,1,2,3,1,2,
+        3,2,1,2,3,2,1,3,  1,3,2,3,1,2,3,1, 2,1,3,1,2,3,2,1,  3,2,1,2,3,1,2,3
+    },
+    
+    {
+        3,3,1,1,3,3,1,1,  1,3,3,1,1,3,3,1, 1,1,3,3,1,1,3,3,  3,1,1,3,3,1,1,3,
+        3,3,1,1,3,3,1,1,  1,3,3,1,1,3,3,1, 1,1,3,3,1,1,3,3,  3,1,1,3,3,1,1,3
+    },
+    
+    {
+        1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2, 1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2,
+        1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2, 1,2,3,2,3,2,1,2,  1,2,3,2,3,2,1,2
+    },
+    
+    {
+        2,2,2,3,2,2,1,2,  2,1,2,2,2,3,2,2, 2,3,2,2,2,1,2,2,  2,2,2,2,2,2,2,1,
+        1,2,2,2,3,2,2,2,  2,2,1,2,2,2,3,2, 2,2,3,2,2,2,1,2,  3,2,2,2,1,2,2,2
+    },
+    
+    {
+        3,2,3,4,2,3,3,2,  3,3,2,3,4,2,3,3, 2,3,4,3,2,3,3,2,  4,2,3,3,2,3,4,2,
+        3,4,2,3,3,2,3,4,  2,3,3,2,3,3,2,3, 3,2,3,3,4,2,3,3,  3,4,2,3,3,2,3,2
+    },
+    
+    {
+        3,3,3,0,3,3,3,3,  3,3,3,0,3,3,3,3, 3,3,3,0,3,3,3,3,  0,0,0,0,0,0,0,0,
+        3,3,3,3,3,3,3,0,  3,3,3,3,3,3,3,0, 3,3,3,3,3,3,3,0,  0,0,0,0,0,0,0,0
+    },
+    
+    {
+        1,1,1,1,1,1,1,1,  2,1,2,2,1,2,2,1, 2,2,1,2,2,1,2,2,  1,1,1,1,1,1,1,1,
+        1,2,2,1,2,2,1,2,  2,1,2,2,1,2,2,1, 2,2,1,2,2,1,2,2,  1,1,1,1,1,1,1,1
+    },
+    
+    {
+        1,1,3,3,3,1,1,2,  2,2,3,3,3,2,2,1, 1,3,3,3,1,1,2,2,  2,3,3,3,2,2,1,1,
+        3,3,3,1,1,2,2,3,  3,3,2,2,1,1,3,3, 1,1,1,2,2,3,3,3,  1,1,2,2,3,3,1,1
+    },
+    
+    {
+        0,4,1,5,0,3,2,5,  4,1,5,0,2,5,1,3, 5,2,0,3,5,1,4,0,  1,5,2,4,0,3,5,2,
+        0,3,5,1,4,0,2,5,  3,0,4,2,5,1,0,4, 4,1,2,5,0,3,2,1,  5,2,0,3,1,4,2,0
+    },
+    
+    {
+        1,2,1,1,1,2,1,1,  1,1,2,1,1,1,2,1, 2,1,1,1,2,1,1,1,  1,1,2,1,1,1,2,1,
+        1,2,1,1,1,2,1,1,  1,1,2,1,1,1,2,1, 2,1,1,1,2,1,1,1,  1,1,2,1,1,1,2,1
+    },
+    
+    {
+        2,1,3,2,1,3,2,1,  3,2,1,2,3,1,2,3, 1,3,2,3,1,2,3,2,  2,1,3,1,2,3,1,2,
+        3,2,1,2,3,2,1,3,  1,3,2,3,1,2,3,1, 2,1,3,1,2,3,2,1,  3,2,1,2,3,1,2,3
+    },
+    
+    {
+        2,3,4,3,2,3,4,3,  3,4,3,2,3,4,3,2, 4,3,2,3,4,3,2,3,  3,2,3,4,3,2,3,4,
+        2,3,4,3,2,3,4,3,  3,4,3,2,3,4,3,2, 4,3,2,3,4,3,2,3,  3,2,3,4,3,2,3,4
+    },
+
+    {
+        1,4,6,4,1,4,6,4,  6,2,0,2,6,4,2,4, 4,0,2,0,4,6,1,6,  4,6,4,6,4,6,4,6,
+        1,4,6,4,1,4,6,4,  6,2,0,2,6,4,2,4, 4,0,2,0,4,6,1,6,  4,6,4,6,4,6,4,6
+    },
+
+   {
+        5,5,5,5,5,5,5,5, 5,0,5,0,5,5,0,5,  1,0,0,1,0,0,0,0, 0,0,0,0,1,0,1,0,
+        1,0,0,1,0,0,1,0, 0,0,0,0,0,1,0,0,  0,1,0,0,1,0,1,0, 1,0,0,1,0,0,0,1
+    }
+};
+
+#endif
